requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.0.0 
gspread==5.12.0
oauth2client==4.1.3
requests==2.31.0
pytz==2024.1